move queryReducer to list controllers directory
