+++ packages/ra-core/src/controller/list/index.ts
@@ -12,3 +12,4 @@ export * from './useListFilterContext';
 export * from './useListPaginationContext';
 export * from './useListParams';
 export * from './useListSortContext';
+export * from './queryReducer';
